Move toppage contents to app dir
