+++ app/toppage/AboutMe.tsx
@@ -1,6 +1,6 @@
 import styles from "../../styles/top-page/AboutMe.module.scss";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import Block from "../Block";
+import Block from "../../components/Block";
 
 import {
   faAt,
