+++ app/toppage/TopPageMusic.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Block from "../Block";
+import Block from "../../components/Block";
 import LiteYouTubeEmbed from "react-lite-youtube-embed";
 import Image from "next/legacy/image";
 import Link from "next/link";
